+++ R/weight-streetnet.R
@@ -20,8 +20,8 @@
 #' (default works with `osmdata` objects).
 #' @param keep_cols Vectors of columns from `x` to be kept in the resultant
 #' `dodgr` network; vector can be either names or indices of desired columns.
-#' @param turn_angle Weight edges times by turning angles at intersections (see
-#' Note).
+#' @param turn_penalty Including time penalty on edges for turning across
+#' oncoming traffic at intersections (see Note).
 #' @param left_side Does traffic travel on the left side of the road (`TRUE`) or
 #' the right side (`FALSE`)? - only has effect on turn angle calculations for
 #' edge times.
@@ -48,11 +48,11 @@
 #' of custom weighting profiles is illustrated in the following example.
 #'
 #' @note Calculating edge times to account for turn angles (that is, with
-#' `turn_angle = TRUE`) involves calculating the temporal delay involving in
+#' `turn_penalty = TRUE`) involves calculating the temporal delay involving in
 #' turning across oncoming traffic. Resultant graphs are fundamentally different
 #' from the default for distance-based routing. The result of
-#' `weight_streetnet(..., turn_angle = TRUE)` should thus \emph{only} be used to
-#' submit to the \link{dodgr_times} function, and not for any other `dodgr`
+#' `weight_streetnet(..., turn_penalty = TRUE)` should thus \emph{only} be used
+#' to submit to the \link{dodgr_times} function, and not for any other `dodgr`
 #' functions nor forms of network analysis.
 #'
 #' @seealso \link{write_dodgr_wt_profile}, \link{dodgr_times}
@@ -91,7 +91,7 @@
 #' }
 weight_streetnet <- function (x, wt_profile = "bicycle",
                               wt_profile_file = NULL,
-                              turn_angle = FALSE,
+                              turn_penalty = FALSE,
                               type_col = "highway", id_col = "osm_id",
                               keep_cols = NULL, left_side = FALSE)
 {
@@ -102,7 +102,7 @@ weight_streetnet <- function (x, wt_profile = "bicycle",
 #' @export
 weight_streetnet.default <- function (x, wt_profile = "bicycle",
                               wt_profile_file = NULL,
-                              turn_angle = FALSE,
+                              turn_penalty = FALSE,
                               type_col = "highway", id_col = "osm_id",
                               keep_cols = NULL, left_side = FALSE)
 {
@@ -124,12 +124,12 @@ way_types_to_keep = c ("highway", "oneway", "oneway.bicycle", "oneway:bicycle",
 #' @export
 weight_streetnet.sf <- function (x, wt_profile = "bicycle",
                                  wt_profile_file = NULL,
-                                 turn_angle = FALSE,
+                                 turn_penalty = FALSE,
                                  type_col = "highway", id_col = "osm_id",
                                  keep_cols = NULL, left_side = FALSE)
 {
-    if (turn_angle)
-        stop ("Turn-angle calculations only currently implemented for street ",
+    if (turn_penalty)
+        stop ("Turn-penalty calculations only currently implemented for street ",
               "network data generated with the `osmdata::osmdata_sc()` function.")
     geom_column <- get_sf_geom_col (x)
     attr (x, "sf_column") <- geom_column
@@ -451,7 +451,7 @@ add_extra_sf_columns <- function (graph, x)
 #' @export
 weight_streetnet.sc <- weight_streetnet.SC <- function (x, wt_profile = "bicycle",
                                                         wt_profile_file = NULL,
-                                                        turn_angle = FALSE,
+                                                        turn_penalty = FALSE,
                                                         type_col = "highway",
                                                         id_col = "osm_id",
                                                         keep_cols = NULL,
@@ -484,7 +484,7 @@ weight_streetnet.sc <- weight_streetnet.SC <- function (x, wt_profile = "bicycle
 
     attr (graph, "turn_penalty") <- 0
 
-    if (turn_angle)
+    if (turn_penalty)
     {
         attr (graph, "turn_penalty") <- 
             get_turn_penalties (wt_profile, wt_profile_file)$turn
@@ -504,7 +504,7 @@ weight_streetnet.sc <- weight_streetnet.SC <- function (x, wt_profile = "bicycle
 
     class (graph) <- c (class (graph), "dodgr_streetnet", "dodgr_streetnet_sc")
 
-    if (turn_angle)
+    if (turn_penalty)
     {
         hash <- digest::digest (graph [[gr_cols$edge_id]])
         fname <- file.path (tempdir (), paste0 ("dodgr_edge_contractions_",